// src/shared/api/models/getApiBundleOffersSoldIdType200Item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiBundleOffersSoldIdType200ItemApplicableServices } from './getApiBundleOffersSoldIdType200ItemApplicableServices';
import type { GetApiBundleOffersSoldIdType200ItemServicesItem } from './getApiBundleOffersSoldIdType200ItemServicesItem';
import type { GetApiBundleOffersSoldIdType200ItemBundleOfferType } from './getApiBundleOffersSoldIdType200ItemBundleOfferType';
import type { GetApiBundleOffersSoldIdType200ItemCustomer } from './getApiBundleOffersSoldIdType200ItemCustomer';
import type { GetApiBundleOffersSoldIdType200ItemEmployee } from './getApiBundleOffersSoldIdType200ItemEmployee';
import type { GetApiBundleOffersSoldIdType200ItemBundleOffer } from './getApiBundleOffersSoldIdType200ItemBundleOffer';

export type GetApiBundleOffersSoldIdType200Item = {
    id?: string;
    bundleOfferId?: string;
    outletId?: number;
    orignalPunches?: number;
    /** @nullable */
    bundleOfferCode?: string | null;
    /** @nullable */
    applicableServices?: GetApiBundleOffersSoldIdType200ItemApplicableServices;
    /**
     * Array of services to use with their quantities
     * @minItems 1
     */
    services?: GetApiBundleOffersSoldIdType200ItemServicesItem[];
    residuePunches?: number;
    status?: string;
    /** @nullable */
    note?: string | null;
    sourceId?: number;
    /** @nullable */
    startDate?: string | null;
    /** @nullable */
    customerId?: number | null;
    price?: number;
    bundleOfferType?: GetApiBundleOffersSoldIdType200ItemBundleOfferType;
    /** @nullable */
    expiryDate?: string | null;
    taxIds?: string[];
    taxAmount?: number;
    /** @nullable */
    personalMessage?: string | null;
    /** @nullable */
    soldBy?: number | null;
    createdAt?: string;
    updatedAt?: string;
    /** @nullable */
    deletedAt?: string | null;
    /** @nullable */
    customer?: GetApiBundleOffersSoldIdType200ItemCustomer;
    /** @nullable */
    employee?: GetApiBundleOffersSoldIdType200ItemEmployee;
    /** @nullable */
    bundleOffer?: GetApiBundleOffersSoldIdType200ItemBundleOffer;
};